Between-species dominance in birds data

Francis et al. (2018) conducted an experiment to study the competitive interactions between 10 species of birds. These are the data presented in Table 1 of this paper.

birds

Format

A data frame with 10 observations on the following 8 variables.

species: a two-letter abbreviation for the bird species.

species_name: the common bird species name.

sci_name: the scientific bird species name.

mass_g: mean species mass in grams.

dom_rank: dominance rank, the smaller the value the more dominant species tends to be over other species.

dom_rank_LCI: lower 95% confidence limit for the dominance rank.

dom_rank_UCI: upper 95% confidence limit for the dominance rank.

Source

S1 dataset. Data for species identities, mass and dominance rankings from Francis et al (2018).

References

Francis M.L., Plummer K.E., Lythgoe B.A., Macallan C., Currie T.E., Blount J.D. (2018) Effects of supplementary feeding on interspecific dominance hierarchies in garden birds. PLoS ONE 13(9): e0202152. https://doi.org/10.1371/journal.pone.0202152