Tokyo 2020 Olympics Men's Individual 10m platform

Detailed results from the final of the Men's Individual 10m platform final at the Tokyo 2020 Olympic Games.

daley

Format

A data frame with 60 observations on the following 17 variables. Each of the 10 competitors has one row of data for each of their 6 dives. For each competitor, the first row relates to the first round of dives and the sixth row the last round.

Rank: Final position (rank) in competition.

Name: Full name.

NOCcode: Country code.

DiveNo: Code for the type of dive performed.

DD: Degree of difficulty: the larger the number the more difficult the dive.

J1-J7: Respective scores from each of the 7 judges.

DivePoints: Points awarded for the dive, based on the scores and the degree of difficulty.

DiveRank: The rank of the dive in the round in question. In the raw data there are entries like "=2", resulting from ties. In daley the "=" sign is removed and the ranks are stored as numeric scalars, e.g. 2.

TotalPoints: The total point accumulated from all dives to this round of the competition.

OverallRank: The rank of the diver after the current round. In the raw data there are entries like "=2", resulting from ties. In daley the "=" sign is removed and the ranks are stored as numeric scalars, e.g. 2.

PointsBehind: The number of points that the diver is behind the leader after the current round. If the diver is the leader then this is missing (NA).

Source

Tokyo 2020 Diving Results, specifically Men's 10m Platform Results.

Details

DivePoints is calculated by: removing the 2 smallest and 2 largest scores from the 7 scores J1-J7; calculating the sum of the 3 scores that remain; and multiplying the result by DD.