Calculation of diving scores

Functions to implement four ways to calculate a score for a dive in a diving competitions, including the way that is used in practice.

dmean(x, DD = "DD", scores = paste0("J", 1:7), trim = 2/7)

dmedian(x, DD = "DD", scores = paste0("J", 1:7), type = 6)

dmode(x, DD = "DD", scores = paste0("J", 1:7))

Arguments

x
A numeric matrix or data frame.

DD
A character (or numeric) scalar giving the column name (or number) of x that contains the DD (degree of difficulty) values of each dive. The default is set up to work with the data in daley1.

scores
A character (or numeric) vector giving the column names (or numbers) of x that contain the 7 scores for each dive. The default is set up to work with the data in daley1.

trim
Only relevant to divemean. Passed to mean. The default, trim = 2 / 7, corresponds to the calculation using in diving competitions.

type
Only relevant to divemedian. Passed to quantile. The default, type = 6, corresponds to the method that we have studied in STAT0002.

Details

In the default case, e.g. dmean(daley1), the 2 smallest and 2 largest scores of the 7 scores are removed; the sum of the 3 remaining scores is calculated; and the result is multiplied by the dive's degree of difficulty.

Another way to think of this is that we take the sample mean of the 3 middle scores, then multiply by 3, then multiply by the degree of difficulty.

In the dmean(daley1, trim = 0) case we take the sample mean of all 7 scores, then multiply by 3, then multiply by the degree of difficulty. That is, we do not trim the 2 smallest and 2 largest values.

Similarly, in the dmedian(daley1) and dmode(daley1) cases we take the sample median or mode, respectively, of all 7 scores, then multiply by 3, then multiply by the degree of difficulty. If there is more than one sample mode then we use the sample mean of these modes.

See also

Examples

# The calculation used in competitions (trimmed sample sums)
dmean(daley1)

#> Error in dmean(daley1): could not find function "dmean"

# An equivalent calculation without trimming
dmean(daley1, trim = 0)

#> Error in dmean(daley1, trim = 0): could not find function "dmean"

# An equivalent calculation based on a sample median
dmedian(daley1)

#> Error in dmedian(daley1): could not find function "dmedian"

# An equivalent calculation based on a sample mode
dmode(daley1)

#> Error in dmode(daley1): could not find function "dmode"