Influenza data

The numbers of people (in thousands of people) in the UK visiting their doctor with symptoms of influenza (flu) during four-weekly time periods over the time period 28th January 1967 to 13th November 2004.

Usage

flu

Format

A data frame with 494 observations on the following 2 variables.

date: the date.

visits: the number of people visiting their doctor with symptoms of flu.