General information about the movies

These movies are animations used to illustrate key ideas in STAT0002. They are produced using the package rpanel.package. You need to install rpanel once, using RStudio's Install button in the Packages menu or install.packages("rpanel") on the command line). Prior to creating a movie the rpanel package must be loaded, using `library(rpanel)`, in order that these functions work. If rpanel is not loaded then an error message similar to Error in shuttle_movie() : could not find function "rp.control" will be produced. For the one of the movies the tkrplot package is also required. See Examples below.

Value

Nothing is returned, only the animation is produced.

Details

(Some of these movies, and other movies are also available in a more user-friendly form, via the smovie package. If you have smovie installed (install.packages("smovie")) then you can access these using library(smovie) and then movies(). See movies for more details.

When one of these functions is called R opens up a small parameter window containing clickable buttons that can be used to change parameters underlying the plot. For the effects of these buttons see the documentation of the individual functions below.

The parameter window does not close automatically after the movie: the user needs to close it manually.

Some movies create objects in the global environment, that is, objects that will be listed when ls() is used. rm can be used to remove these objects if desired. For example rm(name) can be used to remove object name.

Examples

See the examples given for the individual movies listed below.

References

Bowman, A., Crawford, E., Alexander, G. and Bowman, R. W. (2007). rpanel: Simple Interactive Controls for R Functions Using the tcltk Package. Journal of Statistical Software, 17(9), 1-18. http://www.jstatsoft.org/v17/i09/.

See also