Normal p.d.f. movie

A movie to show how the probability density function (p.d.f.) of the normal distribution varies as the mean and/or variance of the distribution are changed.

normal_pdf_movie(
  starting_mean = 0,
  starting_var = 1,
  delta_mean = 0.1,
  delta_var = 0.1
)

Arguments

starting_mean
A numeric scalar. The value of the mean used to produce the first plot of the p.d.f.

starting_var
A positive numeric scalar. The value of the variance used to produce the first plot of the p.d.f.

delta_mean
A numeric scalar. The amount by which the value of the mean is increased/decreased after one click of the +/- button for the mean.

delta_var
A numeric scalar. The amount by which the value of the variance is increased/decreased after one click of the +/- button for the variance.

Value

Nothing is returned, only the animation is produced.

Details

The values of the mean and/or variance can be changed using the +/- buttons in the panel. For the purposes of this movie, the variance cannot be reduced below a value of 0.1.

See also

Examples

# Load package rpanel
# [Use install.packages("rpanel") if necessary]
library(rpanel)

# Produce movie using values from the Aussie births data
if (FALSE) {
normal_pdf_movie()
normal_pdf_movie(delta_var = 1)
}