Oxford Birth Times

Times spent in the delivery suite by 95 women giving birth at the John Radcliffe Hospital, Oxford. These data have been taken from the SMPracticals package to which they were provided by Ethel Burns.

ox_births

Format

A data frame with 95 observations on the following 2 variables.

day: Day on which the woman arrived.

time: Time (hours) spent on delivery suite.

Source

Davison, A. C. (2003) Statistical Models. Cambridge University Press. Page 18.