stat0002: Introduction to Probability and Statistics at UCL

Provides example datasets, R code and tutorials to help students taking STAT0002 Introduction to Probability and Statistics at University College London to understand the course material and to see how R can be used to perform some of the analyses in the course. This package can also be used to complement STAT0004 Introduction to Practical Statistics, enabling students to become more familiar with R code by seeing it in action and playing with it themselves.

Please note

The main sources of information about the material in STAT0002 are the lectures and tutorials, with notes, lecture slides, exercises etc available from the STAT0002 Moodle page. The focus of this R package is mainly on R code.

How to learn about R

The best way to learn how to use a computer program is to play with code and see what it does. R is based on functions that look like name(), where name is the name of the function and where various arguments are inserted inside the brackets (). In the R code provided with this package you will see many functions like this. To find out what a function called name does use either ?name or help(name).

If you have any questions about this package please ask them via the STAT0002 Moodle Discussion Forum.

Vignettes (tutorials demonstrating how to use R code)

Chapter 1: Challenger Space Shuttle Disaster

Chapter 1: Stochastic Simulation

Chapter 2: Descriptive Statistics

Chapter 2: Graphs (One Variable)

Chapter 2: Graphs (More Than One Variable)

Chapter 2: Transformation Of Variable

Datasets

Challenger Space Shuttle Disaster

Mystery data in Exercises 1

T-cell count data in Exercises 2

Student Admissions at UC Berkeley

Oxford Birth Times

Blood Types

FTSE 100 Share Index

World Bank Development Indicators

Influenza data

The 2000 U.S. Presidential Election

Australian Birth Times Data

Nebulae data from Hubble (1929)

Between-species dominance in birds

Movies

Some of the following movies, and other movies, are also available in a more user-friendly form, via the smovie package. If you have smovie installed (install.packages("smovie")) then you can access these using library(smovie) and then movies(). See movies for more details.

Challenger Space Shuttle Disaster

US election: straightening scatter plots

World Bank Development Indicators

Oxford Birth Times Simulation

Binomial p.m.f.

Poisson process

Graphical checking of Poisson process properties

Normal probability density function

Normal central probability areas

Construction of a (normal) QQ plot

Normal sampling distributions

Central Limit Theorem: normal data

Central Limit Theorem: exponential data

Mean vs median for a normal random sample

Mean vs median for a Student's t random sample

Test for lack of association in a 2 by 2 contingency table

Simple linear regression using least squares estimation

Leverage and influence in simple linear regression

Sampling distribution of the correlation coefficient

See also