stat0002: Introduction to Probability and Statistics at UCL

Provides example datasets, R code and tutorials to help students taking STAT0002 Introduction to Probability and Statistics at University College London to understand the course material and to see how R can be used to perform some of the analyses in the course. This package can also be used to complement STAT0004 Introduction to Practical Statistics, enabling students to become more familiar with R code by seeing it in action and playing with it themselves.

Details

The most useful source of information is the stat0002 home page on GitHub. The Reference page provides information about example datasets, movies (interactive plots), and functions for performing tasks like creating plot, calculating summary statistics and simulating data.

The tab labelled Articles (or use the Articles link) contains several tutorials demonstrating how to use R to create some of the content of the STAT0002 notes and related tasks.

Some of the movies are made available via the smovie package. You can access a menu for the movies in smovie using library(smovie) and then movies(). See movies for more details.

Summary of the main resources

Tutorials

Example data

Movies. See also movies

Plots

Summary statistics

Simulation functions

Please note

The main source of information about the material in STAT0002 is the STAT0002 Moodle page. You are not required to use the `stat0002` package during STAT0002, but it is recommended as a supplement to the STAT0002 syllabus.

How to learn about R

The best way to learn how to use a computer program is to play with code and see what it does. R is based on functions that look like name(), where name is the name of the function and where various arguments are inserted inside the brackets (). In the R code provided with this package you will see many functions like this. To find out what a function called name does use either ?name or help(name).

If you have any questions about this package please ask them via the stat0002 R package Moodle Forum.

See also

Author

Maintainer: Paul J. Northrop p.northrop@ucl.ac.uk [copyright holder]