General information about the movies

These movies are animations used to illustrate key ideas in STAT0002.

Value

Nothing is returned, only the animation is produced.

Details

For a list of movies see Movies.

When one of these functions is called R opens up a small parameter window containing clickable buttons that can be used to change parameters underlying the plot. For the effects of these buttons see the documentation of the individual functions below.

The parameter window does not close automatically after the movie: the user needs to close it manually.

Some movies create objects in the global environment, that is, objects that will be listed when ls() is used. rm can be used to remove these objects if desired. For example rm(name) can be used to remove object name.

The smovie package provides a more user-friendly menu from which you can access some of these movies. Use library(smovie) and then movies(). See movies for more details.

Examples

See the examples given for the individual movies detailed in the Movies section of the Reference part of the stat0002 home page on GitHub